+++ flash/core/data/data_module.py
@@ -15,11 +15,9 @@ import os
 import platform
 from typing import Any, Callable, Collection, Dict, Iterable, List, Optional, Sequence, Tuple, TYPE_CHECKING, Union
 
-import networkx as nx
 import numpy as np
 import pytorch_lightning as pl
 import torch
-import torch_geometric
 from pytorch_lightning.trainer.states import RunningStage
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from torch.utils.data import DataLoader, Dataset
@@ -34,7 +32,7 @@ from flash.core.data.data_pipeline import DataPipeline, DefaultPreprocess, Postp
 from flash.core.data.data_source import DataSource, DefaultDataSources
 from flash.core.data.splits import SplitDataset
 from flash.core.data.utils import _STAGES_PREFIX
-from flash.core.utilities.imports import _FIFTYONE_AVAILABLE
+from flash.core.utilities.imports import _FIFTYONE_AVAILABLE, requires
 
 if _FIFTYONE_AVAILABLE and TYPE_CHECKING:
     from fiftyone.core.collections import SampleCollection
@@ -1075,6 +1073,7 @@ class DataModule(pl.LightningDataModule):
         )
 
     @classmethod
+    @requires("fiftyone")
     def from_fiftyone(
         cls,
         train_dataset: Optional[SampleCollection] = None,
@@ -1138,9 +1137,6 @@ class DataModule(pl.LightningDataModule):
                 },
             )
         """
-        if not _FIFTYONE_AVAILABLE:
-            raise ModuleNotFoundError("Please, `pip install fiftyone`.")
-
         return cls.from_data_source(
             DefaultDataSources.FIFTYONE,
             train_dataset,