+++ flash/core/utilities/imports.py
@@ -11,7 +11,6 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""General utilities"""
 import functools
 import importlib
 import operator
@@ -76,14 +75,16 @@ _TORCHVISION_AVAILABLE = _module_available("torchvision")
 _PYTORCHVIDEO_AVAILABLE = _module_available("pytorchvideo")
 _MATPLOTLIB_AVAILABLE = _module_available("matplotlib")
 _TRANSFORMERS_AVAILABLE = _module_available("transformers")
-PYSTICHE_AVAILABLE = _module_available("pystiche")
+_PYSTICHE_AVAILABLE = _module_available("pystiche")
 _FIFTYONE_AVAILABLE = _module_available("fiftyone")
 _FASTAPI_AVAILABLE = _module_available("fastapi")
 _PYDANTIC_AVAILABLE = _module_available("pydantic")
-GRAPHVIZ_AVAILABLE = _module_available("graphviz")
-CYTOOLZ_AVAILABLE = _module_available("cytoolz")
+_GRAPHVIZ_AVAILABLE = _module_available("graphviz")
+_CYTOOLZ_AVAILABLE = _module_available("cytoolz")
 _UVICORN_AVAILABLE = _module_available("uvicorn")
 _PIL_AVAILABLE = _module_available("PIL")
+_ASTEROID_AVAILABLE = _module_available("asteroid")
+_SEGMENTATION_MODELS_AVAILABLE = _module_available("segmentation_models_pytorch")
 _TORCH_SCATTER_AVAILABLE = _module_available("torch_scatter")
 _TORCH_SPARSE_AVAILABLE = _module_available("torch_sparse")
 _TORCH_GEOMETRIC_AVAILABLE = _module_available("torch_geometric")
@@ -99,12 +100,14 @@ _IMAGE_AVAILABLE = all([
     _TIMM_AVAILABLE,
     _PIL_AVAILABLE,
     _KORNIA_AVAILABLE,
-    _MATPLOTLIB_AVAILABLE,
-    _COCO_AVAILABLE,
-    _FIFTYONE_AVAILABLE,
     _PYSTICHE_AVAILABLE,
+    _SEGMENTATION_MODELS_AVAILABLE,
 ])
 _SERVE_AVAILABLE = _FASTAPI_AVAILABLE and _PYDANTIC_AVAILABLE and _CYTOOLZ_AVAILABLE and _UVICORN_AVAILABLE
+_AUDIO_AVAILABLE = all([
+    _ASTEROID_AVAILABLE,
+])
+_GRAPH_AVAILABLE = _TORCH_SCATTER_AVAILABLE and _TORCH_SPARSE_AVAILABLE and _TORCH_GEOMETRIC_AVAILABLE
 
 _GRAPH_AVAILABLE = _TORCH_SCATTER_AVAILABLE and _TORCH_SPARSE_AVAILABLE and _TORCH_GEOMETRIC_AVAILABLE
 
@@ -114,25 +117,35 @@ _EXTRAS_AVAILABLE = {
     'text': _TEXT_AVAILABLE,
     'video': _VIDEO_AVAILABLE,
     'serve': _SERVE_AVAILABLE,
-    'graph': _GRAPH_AVAILABLE
+    'audio': _AUDIO_AVAILABLE,
+    'graph': _GRAPH_AVAILABLE,
 }
 
 
-def _requires_extras(extras: str):
+def _requires(module_path: str, module_available: bool):
 
     def decorator(func):
+        if not module_available:
 
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            if not _EXTRAS_AVAILABLE[extras]:
+            @functools.wraps(func)
+            def wrapper(*args, **kwargs):
                 raise ModuleNotFoundError(
-                    f"Required dependencies not available. Please run: pip install 'lightning-flash[{extras}]'"
+                    f"Required dependencies not available. Please run: pip install '{module_path}'"
                 )
-            return func(*args, **kwargs)
 
-        return wrapper
+            return wrapper
+        else:
+            return func
 
     return decorator
+
+
+def requires(module_path: str):
+    return _requires(module_path, _module_available(module_path))
+
+
+def requires_extras(extras: str):
+    return _requires(f"lightning-flash[{extras}]", _EXTRAS_AVAILABLE[extras])
 
 
 def lazy_import(module_name, callback=None):