+++ flash/graph/classification/data.py
@@ -11,36 +11,30 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Dict, Optional, Sequence
+from typing import Any, Callable, Dict, Optional
 
 from flash.core.data.data_module import DataModule
-from flash.core.data.data_source import DefaultDataKeys, DefaultDataSources
+from flash.core.data.data_source import DefaultDataSources
 from flash.core.data.process import Preprocess
-from flash.core.data.transforms import ApplyToKeys
-from flash.core.utilities.imports import _PYTORCH_GEOMETRIC_AVAILABLE
-from flash.graph.data_source import GraphDatasetSource
+from flash.core.utilities.imports import _GRAPH_AVAILABLE, requires_extras
+from flash.graph.data import GraphDatasetDataSource
 
-if _PYTORCH_GEOMETRIC_AVAILABLE:
-    import networkx as nx
-    from torch_geometric.data import DataLoader, Dataset
+if _GRAPH_AVAILABLE:
     from torch_geometric.data.batch import Batch
-
-# See https://1176-333857397-gh.circle-artifacts.com/0/html/task_template.html
 
 
 class GraphClassificationPreprocess(Preprocess):
 
+    @requires_extras("graph")
     def __init__(
         self,
         train_transform: Optional[Dict[str, Callable]] = None,
         val_transform: Optional[Dict[str, Callable]] = None,
         test_transform: Optional[Dict[str, Callable]] = None,
         predict_transform: Optional[Dict[str, Callable]] = None,
-        num_features: int = 128  #todo: do we want to add backbone here as in text?
+        num_features: int = 128,
     ):
         self.num_features = num_features
-        if not _PYTORCH_GEOMETRIC_AVAILABLE:
-            raise ModuleNotFoundError("Please, pip install -e '.[graph]'")
 
         super().__init__(
             train_transform=train_transform,
@@ -48,7 +42,7 @@ class GraphClassificationPreprocess(Preprocess):
             test_transform=test_transform,
             predict_transform=predict_transform,
             data_sources={
-                DefaultDataSources.DATASET: GraphDatasetSource(),
+                DefaultDataSources.DATASET: GraphDatasetDataSource(),
             },
             default_data_source=DefaultDataSources.DATASET,
         )
@@ -60,8 +54,9 @@ class GraphClassificationPreprocess(Preprocess):
     def load_state_dict(cls, state_dict: Dict[str, Any], strict: bool = False):
         return cls(**state_dict)
 
-    def collate(self, samples: Sequence) -> Any:
-        return {DefaultDataKeys.INPUT: Batch.from_data_list([s[DefaultDataKeys.INPUT] for s in samples])}
+    @staticmethod
+    def default_transforms() -> Optional[Dict[str, Callable]]:
+        return {"collate": Batch.from_data_list}
 
 
 class GraphClassificationData(DataModule):