+++ flash/graph/classification/model.py
@@ -14,7 +14,6 @@
 from types import FunctionType
 from typing import Any, Callable, List, Mapping, Optional, Sequence, Tuple, Type, Union, Dict
 
-import pytorch_lightning as pl
 import torch
 from pytorch_lightning.metrics import Accuracy
 from torch import nn
@@ -26,9 +25,9 @@ from flash.core.data.data_source import DefaultDataKeys
 from flash.core.data.process import Serializer
 from flash.core.registry import FlashRegistry
 from flash.graph.backbones import GRAPH_CLASSIFICATION_BACKBONES
-from flash.core.utilities.imports import _PYTORCH_GEOMETRIC_AVAILABLE
+from flash.core.utilities.imports import _TORCH_GEOMETRIC_AVAILABLE
 
-if _PYTORCH_GEOMETRIC_AVAILABLE:
+if _TORCH_GEOMETRIC_AVAILABLE:
     from torch_geometric.nn import BatchNorm, GCNConv, global_mean_pool, MessagePassing
 else:
     MessagePassing = type(object)