+++ flash/text/seq2seq/core/data.py
@@ -23,7 +23,7 @@ from flash.core.data.data_module import DataModule
 from flash.core.data.data_source import DataSource, DefaultDataSources
 from flash.core.data.process import Postprocess, Preprocess
 from flash.core.data.properties import ProcessState
-from flash.core.utilities.imports import _requires_extras, _TEXT_AVAILABLE
+from flash.core.utilities.imports import _TEXT_AVAILABLE, requires_extras
 from flash.text.classification.data import TextDeserializer
 
 if _TEXT_AVAILABLE:
@@ -34,7 +34,7 @@ if _TEXT_AVAILABLE:
 
 class Seq2SeqDataSource(DataSource):
 
-    @_requires_extras("text")
+    @requires_extras("text")
     def __init__(
         self,
         backbone: str,
@@ -218,7 +218,7 @@ class Seq2SeqBackboneState(ProcessState):
 
 class Seq2SeqPreprocess(Preprocess):
 
-    @_requires_extras("text")
+    @requires_extras("text")
     def __init__(
         self,
         train_transform: Optional[Dict[str, Callable]] = None,
@@ -286,7 +286,7 @@ class Seq2SeqPreprocess(Preprocess):
 
 class Seq2SeqPostprocess(Postprocess):
 
-    @_requires_extras("text")
+    @requires_extras("text")
     def __init__(self):
         super().__init__()