+++ tests/graph/classification/test_data.py
@@ -11,29 +11,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-import random
-from pathlib import Path
-
-import networkx as nx
 import pytest
-from torch.functional import Tensor
 
-from flash.core.data.data_source import DefaultDataKeys
-from flash.core.utilities.imports import _PYTORCH_GEOMETRIC_AVAILABLE
+from flash.core.data.transforms import merge_transforms
+from flash.core.utilities.imports import _TORCH_GEOMETRIC_AVAILABLE
 from flash.graph.classification.data import GraphClassificationData, GraphClassificationPreprocess
+from tests.helpers.utils import _GRAPH_TESTING
 
-if _PYTORCH_GEOMETRIC_AVAILABLE:
-    import torch_geometric
-    from torch_geometric.data import data as PyGData
-    from torch_geometric.data import Dataset, download_url
+if _TORCH_GEOMETRIC_AVAILABLE:
     from torch_geometric.datasets import TUDataset
     from torch_geometric.transforms import OneHotDegree
 
 
-@pytest.mark.skipif(not _PYTORCH_GEOMETRIC_AVAILABLE, reason="pytorch geometric isn't installed.")
-class TestTemplatePreprocess:
-    """Tests ``TemplatePreprocess``."""
+@pytest.mark.skipif(not _GRAPH_TESTING, reason="graph libraries aren't installed.")
+class TestGraphClassificationPreprocess:
+    """Tests ``GraphClassificationPreprocess``."""
 
     def test_smoke(self):
         """A simple test that the class can be instantiated."""
@@ -41,24 +33,20 @@ class TestTemplatePreprocess:
         assert prep is not None
 
 
-@pytest.mark.skipif(not _PYTORCH_GEOMETRIC_AVAILABLE, reason="pytorch geometric isn't installed")
+@pytest.mark.skipif(not _GRAPH_TESTING, reason="graph libraries aren't installed.")
 class TestGraphClassificationData:
-    """Tests ``estGraphClassificationData``."""
+    """Tests ``GraphClassificationData``."""
 
     def test_smoke(self):
-        """A simple test that the class can be instantiated."""
         dm = GraphClassificationData()
         assert dm is not None
 
     def test_from_datasets(self, tmpdir):
-        tmpdir = Path(tmpdir)
-        tudataset = TUDataset(root='tmpdir', name='KKI')
-        """Tests that ``TemplateData`` is properly created when using the ``from_dataset`` method."""
-        train_dataset = tudataset,
-        val_dataset = tudataset,
-        test_dataset = tudataset,
-        val_dataset = tudataset,
-        predict_dataset = tudataset,
+        tudataset = TUDataset(root=tmpdir, name='KKI')
+        train_dataset = tudataset
+        val_dataset = tudataset
+        test_dataset = tudataset
+        predict_dataset = tudataset
 
         # instantiate the data module
         dm = GraphClassificationData.from_datasets(
@@ -69,7 +57,8 @@ class TestGraphClassificationData:
             train_transform=None,
             val_transform=None,
             test_transform=None,
-            predict_transform=None
+            predict_transform=None,
+            batch_size=2
         )
         assert dm is not None
         assert dm.train_dataloader() is not None
@@ -78,31 +67,25 @@ class TestGraphClassificationData:
 
         # check training data
         data = next(iter(dm.train_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == tudataset.num_features
-        assert list(targets.size()) == [1]
+        assert list(data.x.size())[1] == tudataset.num_features
+        assert list(data.y.size()) == [2]
 
         # check val data
         data = next(iter(dm.val_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == tudataset.num_features
-        assert list(targets.size()) == [1]
+        assert list(data.x.size())[1] == tudataset.num_features
+        assert list(data.y.size()) == [2]
 
         # check test data
         data = next(iter(dm.test_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == tudataset.num_features
-        assert list(targets.size()) == [1]
+        assert list(data.x.size())[1] == tudataset.num_features
+        assert list(data.y.size()) == [2]
 
     def test_transforms(self, tmpdir):
-        tmpdir = Path(tmpdir)
-        tudataset = TUDataset(root='tmpdir', name='KKI')
-        """Tests that ``TemplateData`` is properly created when using the ``from_dataset`` method."""
-        train_dataset = tudataset,
-        val_dataset = tudataset,
-        test_dataset = tudataset,
-        val_dataset = tudataset,
-        predict_dataset = tudataset,
+        tudataset = TUDataset(root=tmpdir, name='KKI')
+        train_dataset = tudataset
+        val_dataset = tudataset
+        test_dataset = tudataset
+        predict_dataset = tudataset
 
         # instantiate the data module
         dm = GraphClassificationData.from_datasets(
@@ -110,10 +93,23 @@ class TestGraphClassificationData:
             val_dataset=val_dataset,
             test_dataset=test_dataset,
             predict_dataset=predict_dataset,
-            train_transform=OneHotDegree,
-            val_transform=OneHotDegree,
-            test_transform=OneHotDegree,
-            predict_transform=OneHotDegree
+            train_transform=merge_transforms(
+                GraphClassificationPreprocess.default_transforms(),
+                {"pre_tensor_transform": OneHotDegree(tudataset.num_features - 1)},
+            ),
+            val_transform=merge_transforms(
+                GraphClassificationPreprocess.default_transforms(),
+                {"pre_tensor_transform": OneHotDegree(tudataset.num_features - 1)},
+            ),
+            test_transform=merge_transforms(
+                GraphClassificationPreprocess.default_transforms(),
+                {"pre_tensor_transform": OneHotDegree(tudataset.num_features - 1)},
+            ),
+            predict_transform=merge_transforms(
+                GraphClassificationPreprocess.default_transforms(),
+                {"pre_tensor_transform": OneHotDegree(tudataset.num_features - 1)},
+            ),
+            batch_size=2,
         )
         assert dm is not None
         assert dm.train_dataloader() is not None
@@ -122,46 +118,15 @@ class TestGraphClassificationData:
 
         # check training data
         data = next(iter(dm.train_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == tudataset.num_features
-        assert list(targets.size()) == [1]
+        assert list(data.x.size())[1] == tudataset.num_features * 2
+        assert list(data.y.size()) == [2]
 
         # check val data
         data = next(iter(dm.val_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == tudataset.num_features
-        assert list(targets.size()) == [1]
+        assert list(data.x.size())[1] == tudataset.num_features * 2
+        assert list(data.y.size()) == [2]
 
         # check test data
         data = next(iter(dm.test_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == tudataset.num_features
-        assert list(targets.size()) == [1]
-
-    def test_from_pygdatasequence(self):
-
-        # instantiate the data module
-        dm = GraphClassificationData.from_pygdatasequence(
-            train_data=[
-                torch_geometric.utils.from_networkx(
-                    nx.complete_bipartite_graph(random.randint(1, 10), random.randint(1, 10))
-                ),
-                torch_geometric.utils.from_networkx(nx.tetrahedral_graph()),
-                torch_geometric.utils.from_networkx(
-                    nx.complete_bipartite_graph(random.randint(1, 10), random.randint(1, 10))
-                ),
-            ],
-            train_targets=[random.randint(0, 1), 0, random.randint(0, 1)],
-            train_transform={
-                torch_geometric.transforms.Cartesian(),
-            },
-        )
-
-        assert dm is not None
-        assert dm.train_dataloader() is not None
-
-        # check training data
-        data = next(iter(dm.train_dataloader()))
-        input, targets = data[DefaultDataKeys.INPUT], data[DefaultDataKeys.TARGET]
-        assert list(input.size())[1] == 1
-        assert list(targets.size()) == [1]
+        assert list(data.x.size())[1] == tudataset.num_features * 2
+        assert list(data.y.size()) == [2]